:recycle: 닉네임 중복 체크 메서드를 UserWorker로 분리
