+++ iOS/Layover/Layover/Scenes/SignUpScene/SignUpWorker.swift
@@ -10,7 +10,6 @@ import Foundation
 import OSLog
 
 protocol SignUpWorkerProtocol {
-    func isDuplicate(username: String) async -> Bool
     func signUp(withKakao socialToken: String, username: String) async -> Bool
     func signUp(withApple identityToken: String, username: String) async -> Bool
 }
@@ -20,18 +19,15 @@ final class SignUpWorker {
     // MARK: Properties
 
     private let signUpEndPointFactory: SignUpEndPointFactory
-    private let userEndPointFactory: UserEndPointFactory
     private let provider: ProviderType
     private let authManager: AuthManagerProtocol
 
     // MARK: Intializer
 
     init(signUpEndPointFactory: SignUpEndPointFactory = DefaultSignUpEndPointFactory(),
-         userEndPointFactory: UserEndPointFactory = DefaultUserEndPointFactory(),
          provider: ProviderType = Provider(),
          authManager: AuthManagerProtocol = AuthManager.shared) {
         self.signUpEndPointFactory = signUpEndPointFactory
-        self.userEndPointFactory = userEndPointFactory
         self.provider = provider
         self.authManager = authManager
     }
@@ -40,20 +36,6 @@ final class SignUpWorker {
 // MARK: - SignUpWorkerProtocol
 
 extension SignUpWorker: SignUpWorkerProtocol {
-    func isDuplicate(username: String) async -> Bool {
-        let endPoint = userEndPointFactory.makeCheckUserNameEndPoint(username: username)
-        do {
-            let responseData = try await provider.request(with: endPoint, authenticationIfNeeded: false)
-            guard let data = try await provider.request(with: endPoint, authenticationIfNeeded: false).data else {
-                os_log(.error, log: .default, "Failed to check duplicate username with error: %@", responseData.message)
-                return true
-            }
-            return data.exist
-        } catch {
-            os_log(.error, log: .default, "Failed to check duplicate username with error: %@", error.localizedDescription)
-            return true
-        }
-    }
 
     func signUp(withKakao socialToken: String, username: String) async -> Bool {
         let endPoint = signUpEndPointFactory.makeKakaoSignUpEndPoint(socialToken: socialToken, username: username)
